doing some fixings

+++ routes/web.php
@@ -6,15 +6,16 @@ use App\Http\Controllers\RoomController;
 use App\Http\Controllers\MessageController;
 use App\Http\Controllers\ReservationController;
 use App\Http\Middleware\checkAdminMiddlware;
+use App\Http\Controllers\DashboardController;
 use App\Models\Room;
 
 Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::get('/dashboard', [DashboardController::class, 'getUserDashboard'])->middleware(['auth', 'verified'])->name('user.dashboard');
+
+Route::get('/admin/dashboard', [DashboardController::class, 'getAdminDashboard'])->middleware(['auth', 'verified',checkAdminMiddlware::class])->name('admin.dashboard');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
@@ -25,15 +26,17 @@ Route::middleware('auth')->group(function () {
 
 /// Rooms routes
 Route::get('/', [RoomController::class, 'index'])->name('room.index');
-Route::get('/room/{id}', [RoomController::class, 'show'])->name('room.show');
+Route::get('/room/{room}', [RoomController::class, 'show'])->name('room.show');
 
-Route::get('/dashboard/room/create', [RoomController::class, 'create'])->middleware(['auth',  checkAdminMiddlware::class])->name('room.create');
 
-Route::get('/dashboard/rooms', [Roomcontroller::class, 'getAll'])->middleware(['auth', checkAdminMiddlware::class])->name('room.all');
 
-Route::post('/dashboard/room/store', [Roomcontroller::class, 'store'])->middleware(['auth', checkAdminMiddlware::class])->name('room.store');
+Route::get('/admin/dashboard/room/create', [RoomController::class, 'create'])->middleware(['auth',  checkAdminMiddlware::class])->name('room.create');
 
-Route::delete('/dashboard/room/delete/{room}', [RoomController::class, 'delete'])->middleware(['auth', checkAdminMiddlware::class])->name('room.delete');
+Route::get('/admin/dashboard/rooms', [Roomcontroller::class, 'getAll'])->middleware(['auth', checkAdminMiddlware::class])->name('room.all');
+
+Route::post('/admin/dashboard/room/store', [Roomcontroller::class, 'store'])->middleware(['auth', checkAdminMiddlware::class])->name('room.store');
+
+Route::delete('/admin/dashboard/room/delete/{room}', [RoomController::class, 'delete'])->middleware(['auth', checkAdminMiddlware::class])->name('room.delete');
 require __DIR__.'/auth.php';
 
 // Messages routes
@@ -42,4 +45,4 @@ Route::post('/message/send', [MessageController::class, 'send'])->name('message.
 // Reservation routes
 Route::post('/rooms/reservation/{room}', [ReservationController::class, 'make'])->middleware(['auth'])->name("reservation.make");
 
-Route::get('/dashboard/reservations', [ReservationController::class, "index"])->middleware(['auth', checkAdminMiddlware::class])->name("reservation.index");
+Route::get('/admin/dashboard/reservations', [ReservationController::class, "index"])->middleware(['auth', checkAdminMiddlware::class])->name("reservation.index");
